refactor: split cli runner from library exports
